Port landing page onto new architecture
Deployed on arnavrawat.xyz

TODO: Port bandwidth monitor to new architecture, taking advantage
of the TLS socket abstraction for websocket.
Make sure we are cleaning up after ourselves by consuming finished
child tasks.
Use delayed deletion to prevent any possible use-after-frees.
Move the http client used for lennox accesses to use io_uring also, and
create a client-type socket adaptor class for TLS for it.
Fix the climate status, the setpoint is wrong.

+++ src/common/inet_socket_read_write_event_bytes.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+#include <iostream>
 #include "io_event.hpp"
 #include "io_uring_manager.hpp"
 #include "defs.hpp"
@@ -27,7 +29,7 @@ public:
         return {"Write len bytes from buf", true, OpHint::OP_HINT_WRITE | OpHint::OP_HINT_NETWORK};
     }
 
-    std::pair<GetDataInfo, void*> get_data(int id) {
+    std::pair<GetDataInfo, void*> get_data(uint64_t id) {
         if (id == ID_READ) {
             return {{true}, read_buffer_};
         }
